+++ rules/node.js
@@ -1,1 +1,3 @@
-module.exports = {};
+module.exports = {
+  'global-require': 'off', // We don't develop for Node environment.
+};